% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\name{get_resources_info}
\alias{get_resources_info}
\title{Get info about resources for a country}
\usage{
get_resources_info(iso3)
}
\arguments{
\item{iso3}{Country iso3 to get resources info for}
}
\value{
A list containing info about resources for this country
}
\description{
This returns a list containing
\itemize{
\item The version of the package installed
\item The URL from the DESCRIPTION
\item A data frame with the files and their hashes for this country
}
}
