% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{get_shipp_workbook_path}
\alias{get_shipp_workbook_path}
\title{Get the path to the workbook template}
\usage{
get_shipp_workbook_path()
}
\value{
Path to the workbook template
}
\description{
Get the path to the workbook template
}
