% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{get_srb_year}
\alias{get_srb_year}
\title{Get year of SRB estimates from SHIPP workbook template}
\usage{
get_srb_year(iso3)
}
\arguments{
\item{iso3}{Country iso3}
}
\value{
Year of SRB estimates to use in SHIPP calculations. Based on most recent
survey or set to 2018 is most recent survey is older than 2018. Harcoded into
SHIPP workbook template on "County Model" tab.
}
\description{
Get year of SRB estimates from SHIPP workbook template
}
