% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw.R
\name{get_workbook_template_path}
\alias{get_workbook_template_path}
\title{Get the path to the workbook template}
\usage{
get_workbook_template_path()
}
\value{
Path to the workbook template
}
\description{
Get the path to the workbook template
}
