% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw.R
\name{load_agyw_exdata}
\alias{load_agyw_exdata}
\title{Load dependencies for AGYW PSE tool}
\usage{
load_agyw_exdata(data, iso3)
}
\arguments{
\item{data}{Name of data dependency (see below for full description).}

\item{iso3}{Country iso3}
}
\value{
Data input for AGYW PSE tool.
}
\description{
Load dependencies for AGYW PSE tool
}
