% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agyw.R
\name{get_agyw_workbook_path}
\alias{get_agyw_workbook_path}
\title{Get the path to the workbook template}
\usage{
get_agyw_workbook_path(iso3)
}
\arguments{
\item{iso3}{Country iso3}
}
\value{
Path to the workbook template
}
\description{
Get the path to the workbook template
}
