% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{load_shipp_exdata}
\alias{load_shipp_exdata}
\title{Load dependencies for SHIPP PSE tool}
\usage{
load_shipp_exdata(data, iso3)
}
\arguments{
\item{data}{Name of data dependency (see below for full description).}

\item{iso3}{Country iso3}
}
\value{
Data input for SHIPP PSE tool.

Loaded SHIPP data.
}
\description{
Load dependencies for SHIPP PSE tool
}
